Move DMD into data_models
