+++ src/prog_models/data_models/__init__.py
@@ -3,5 +3,12 @@
 
 from .data_model import DataModel
 from .lstm_model import LSTMStateTransitionModel
+from .dmd import SurrogateDMDModel
+
+SURROAGATE_METHOD_LOOKUP = {
+    'dmd': SurrogateDMDModel,
+    'lstm': LSTMStateTransitionModel
+}
+
 
 __all__ = ['DataModel', 'LSTMStateTransitionModel']
