--- src/prog_models/surrogates/__init__.py
@@ -1,8 +1,0 @@
-# Copyright © 2021 United States Government as represented by the Administrator of the
-# National Aeronautics and Space Administration.  All Rights Reserved.
-
-from .dmd import SurrogateDMDModel
-
-SURROAGATE_METHOD_LOOKUP = {
-    'dmd': SurrogateDMDModel
-}
